+++ core/static/assets/gulpfile.js
@@ -1,74 +1,30 @@
 /*
 
 =========================================================
-* Volt Free - Bootstrap 5 Dashboard
+* AppSeed - Simple SCSS compiler via Gulp
 =========================================================
-
-* Product Page: https://themesberg.com/product/admin-dashboard/volt-premium-bootstrap-5-dashboard
-* Copyright 2020 Themesberg (https://www.themesberg.com)
-* License (https://themesberg.com/licensing)
-
-* Designed and coded by https://themesberg.com
-
-=========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software. Please contact us to request a removal.
 
 */
 
 var autoprefixer = require('gulp-autoprefixer');
 var browserSync = require('browser-sync').create();
 var cleanCss = require('gulp-clean-css');
-var del = require('del');
-const htmlmin = require('gulp-htmlmin');
-const cssbeautify = require('gulp-cssbeautify');
 var gulp = require('gulp');
 const npmDist = require('gulp-npm-dist');
 var sass = require('gulp-sass')(require('node-sass'));
 var wait = require('gulp-wait');
 var sourcemaps = require('gulp-sourcemaps');
-var fileinclude = require('gulp-file-include');
 var rename = require("gulp-rename");
 
-// Define paths
+// Define COMMON paths
 
 const paths = {
-    dist: {
-        base: './dist/',
-        css: './dist/css',
-        html: './dist/pages',
-        assets: './dist/assets',
-        img: './dist/assets/img',
-        vendor: './dist/vendor'
-    },
-    dev: {
-        base: './html&css/',
-        css: './html&css/css',
-        html: './html&css/pages',
-        assets: './html&css/assets',
-        img: './html&css/assets/img',
-        vendor: './html&css/vendor'
-    },
-    base: {
-        base: './',
-        node: './node_modules'
-    },
     src: {
         base: './',
         css: './css',
-        html: './src/pages/**/*.html',
-        assets: './src/assets/**/*.*',
-        partials: './src/partials/**/*.html',
         scss: './scss',
         node_modules: './node_modules/',
         vendor: './vendor'
-    },
-    temp: {
-        base: './.temp/',
-        css: './.temp/css',
-        html: './.temp/pages',
-        assets: './.temp/assets',
-        vendor: './.temp/vendor'
     }
 };
 
@@ -99,5 +55,5 @@ gulp.task('minify:css', function() {
         .pipe(gulp.dest(paths.src.css))
 });
 
-// Default
+// Default Task: Compile SCSS and minify the result
 gulp.task('default', gulp.series('scss', 'minify:css'));